Renaming min_iteration2 to results_screen, and moving over updated groupformer listings, urls, etc
