--- min_iteration2/templates/main/groupformer_list.html
@@ -1,102 +1,0 @@
-{% load static %}
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        
-        <!-- Custom CSS file -->
-        <link rel="stylesheet" type="text/css" href="{% static 'main/style.css' %}">
-        <!-- Bootstrap CDN -->
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
-        
-        <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
-
-        <title>Sample List Of Groupformer Instances</title>
-    </head>
-
-    <body>
-        <div id="page-container">
-            <div id="container-fluid">
-                <!-- Page Header -->
-                <div class="row">
-                    <h1 class="display-4 text-center mx-auto d-block">Sample List Of Groupformer Instances</h1>
-                </div>
-
-                <!-- Page Spacer -->
-                <div class="spacer"></div>
-
-                <!-- Main Content -->
-                <div class="row">
-                    <main role="main" class="col-8 mx-auto px-4">
-
-                        <!-- Project Listings -->
-                        {% if groupformers %}
-                            {% for groupformer in groupformers %}
-                            <div class="custom-form-box">
-                                <form action='#' class="needs-validation" novalidate>
-                                        <form class="form-inline" action="#">
-                                            <h4>{{ groupformer.section }}</h4>
-                                            <input type="hidden" name="id" value="{{ groupformer.id }}">
-                                            <button type="submit" id="groupformer{{ groupformer.id }}_submit" class="btn btn-primary">Create Groups</button>
-                                        </form>
-
-                                        <!-- Group Names display -->
-                                        <div id="groupformer{{ groupformer.id }}_groups">
-                                        
-                                        </div>
-                                </form>
-                            </div>
-                            <br>
-                            {% endfor %}
-                        {% endif %}
-                    </main>
-                </div> <!-- End Main Content row -->
-            </div>
-        </div>
-
-        <script>
-            
-            function updateGroupDisplay(display, data) {
-                // Create an unordered list of student groups and add them to the display div.
-                var label = $('<h6/>').html("Groups:");
-                var ul = $('<ul/>');
-                ul.addClass('list-group list-group-flush');
-                $.each(data, function (index, value) {
-                    var row = $('<li/>').html(value.join(", "));
-                    row.addClass('list-group-item');
-                    ul.append(row);
-                });
-
-                // Clear display div first to prevent duplicated output.
-                display.html("<br>");
-                display.append(label);
-                display.append(ul);
-            }
-
-            // When any form is submitted, grab the groups formed at the specified form's groupformer id.
-            $("form").submit(function (event) {
-                // Do not actually "submit" the form.
-                event.preventDefault();
-
-                // Save the GroupFormer section ID and select the group display div.
-                var id = this.id.value;
-                var group_display = $("#groupformer" + id + "_groups");
-
-                // Retrieve the created Groups (EXAMPLE FOR PROOF OF CONCEPT)
-                $.ajax({
-                    url: 'sample_groups/' + id,
-                    type: 'get',
-                    dataType: 'json',
-                    success: function (response) {
-                        // Data is in a "data" key, to serialize information as JSON. Cannot send plain List/Arrays over.
-                        data = response.data;
-                        updateGroupDisplay(group_display, data);
-                    }
-                });
-            });
-        </script>
-        <!-- Bootstrap Scripts (STAY AT BOTTOM) -->
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/js/bootstrap.bundle.min.js" integrity="sha384-JEW9xMcG8R+pH31jmWH6WWP0WintQrMb4s7ZOdauHnUtxwoG2vI5DkLtS3qm9Ekf" crossorigin="anonymous"></script>
-    </body>
-
-</html>
