+++ response_screen/views.py
@@ -21,42 +21,6 @@ def response_screen(request, groupformer_id):
     return HttpResponse("404", status=404)
 
 
-def groupformer_list(request):
-    # Get all groupformers to list groupformer instances for group generating
-    groupformers = GroupFormer.objects.all()
-    context = {"groupformers": groupformers}
-
-    return render(request, 'response_screen_main/groupformer_list.html', context)
-
-
-def sample_groups(request, groupformer_id):
-    # Temporary "API endpoint" for retrieving groups
-    # Create arbitrary sample groups for testing the front-end
-    sections = {}
-
-    # Even ids
-    groups = []
-    groups.append(["A","B","C"])
-    groups.append(["1","2","3"])
-    groups.append(["X","Y","Z"])
-    
-    sections[0] = groups
-
-    # Odd ids
-    groups = []
-    groups.append(["Q","A","Z"])
-    groups.append(["G","M","E"])
-    groups.append(["A","S","D","F"])
-
-    sections[1] = groups
-
-    # Subtract 1 to group former ID because ABC is on even ids, and PKs count from 1
-    if (groupformer_id - 1) % 2 in sections:
-        return JsonResponse({"data":sections[(groupformer_id - 1) % 2]})
-    else:
-        return JsonResponse({"data":[]}, status=404)
-
-
 def login_group(request, groupformer_id):
     gfs = GroupFormer.objects.filter(pk=groupformer_id)
     if len(gfs) == 0:
