+++ results_screen/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class MinIteration2Config(AppConfig):
-    name = 'min_iteration2'
+    name = 'results_screen'
